ai-service/requirements.txt
tensorflow>=2.10.0
numpy>=1.21.0
pillow>=9.0.0
opencv-python>=4.5.0
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.12.0
pyyaml>=6.0.0
